requirements.txt
pandas
numpy
scikit-learn
mlflow
joblib
pyyaml
fastapi
uvicorn
pytest
